refactor(dedicated.server): move ftpbackupaccess to bm-server-component
ref: MANAGER-12287
