+++ packages/manager/modules/bm-server-components/src/ftp-backup-storage/access/add/dedicated-server-ftp-backup-access-add.component.js
@@ -1,0 +1,11 @@
+import controller from './dedicated-server-ftp-backup-access-add.controller';
+import template from './dedicated-server-ftp-backup-access-add.html';
+
+export default {
+  controller,
+  template,
+  bindings: {
+    goBack: '<',
+    results: '<',
+  },
+};
