+++ packages/manager/modules/bm-server-components/src/ftp-backup-storage/access/add/dedicated-server-ftp-backup-access-add.html
@@ -1,7 +1,7 @@
-<div data-ng-controller="AddAccessFtpBackupCtrl as $ctrl">
+<div>
     <div
         data-wizard
-        data-wizard-on-cancel="$ctrl.$scope.resetAction"
+        data-wizard-on-cancel="$ctrl.goBack.bind($ctrl)"
         data-wizard-on-finish="$ctrl.addFtpBackup()"
         data-wizard-title=":: 'server_configuration_ftpbackup_add' | translate"
         data-wizard-confirm-button-text=":: 'common_terminate' | translate"
