+++ packages/manager/modules/bm-server-components/src/ftp-backup-storage/access/add/dedicated-server-ftp-backup-access-add.module.js
@@ -1,0 +1,10 @@
+import component from './dedicated-server-ftp-backup-access-add.component';
+
+const moduleName = 'ovhManagerDedicatedServerFtpBackupAccessAdd';
+
+angular
+  .module(moduleName, ['oui', 'pascalprecht.translate'])
+  .component('dedicatedServerFtpBackupAccessAdd', component)
+  .run(/* @ngTranslationsInject:json ./translations */);
+
+export default moduleName;
