+++ packages/manager/modules/bm-server-components/src/ftp-backup-storage/access/delete/dedicated-server-ftp-backup-access-delete.component.js
@@ -1,0 +1,11 @@
+import controller from './dedicated-server-ftp-backup-access-delete.controller';
+import template from './dedicated-server-ftp-backup-access-delete.html';
+
+export default {
+  controller,
+  template,
+  bindings: {
+    goBack: '<',
+    ipbackup: '<',
+  },
+};
