+++ packages/manager/modules/bm-server-components/src/ftp-backup-storage/access/delete/dedicated-server-ftp-backup-access-delete.html
@@ -1,12 +1,10 @@
-<div
-    data-ng-controller="DedicatedServerFtpBackupAccessDeleteController as $ctrl"
->
+<div>
     <oui-modal
         data-primary-action="$ctrl.deleteFtpBackupAccess()"
         data-primary-label="{{:: 'wizard_confirm' | translate }}"
-        data-secondary-action="resetAction()"
+        data-secondary-action="$ctrl.goBack()"
         data-secondary-label="{{:: 'wizard_cancel' | translate }}"
-        data-on-dismiss="resetAction()"
+        data-on-dismiss="$ctrl.goBack()"
         data-loading="$ctrl.isDeleting"
         data-heading="{{:: 'server_configuration_ftpbackup_access_delete' | translate }}"
     >
