+++ packages/manager/modules/bm-server-components/src/ftp-backup-storage/access/delete/dedicated-server-ftp-backup-access-delete.module.js
@@ -1,0 +1,10 @@
+import component from './dedicated-server-ftp-backup-access-delete.component';
+
+const moduleName = 'ovhManagerDedicatedServerFtpBackupAccessDelete';
+
+angular
+  .module(moduleName, ['oui', 'pascalprecht.translate'])
+  .component('dedicatedServerFtpBackupAccessDelete', component)
+  .run(/* @ngTranslationsInject:json ./translations */);
+
+export default moduleName;
