+++ packages/manager/modules/bm-server-components/src/ftp-backup-storage/access/update/dedicated-server-ftp-backup-access-update.component.js
@@ -1,0 +1,11 @@
+import controller from './dedicated-server-ftp-backup-access-update.controller';
+import template from './dedicated-server-ftp-backup-access-update.html';
+
+export default {
+  controller,
+  template,
+  bindings: {
+    goBack: '<',
+    ipbackup: '<',
+  },
+};
