+++ packages/manager/modules/bm-server-components/src/ftp-backup-storage/access/update/dedicated-server-ftp-backup-access-update.module.js
@@ -1,0 +1,10 @@
+import component from './dedicated-server-ftp-backup-access-update.component';
+
+const moduleName = 'ovhManagerDedicatedServerFtpBackupAccessUpdate';
+
+angular
+  .module(moduleName, ['oui', 'pascalprecht.translate'])
+  .component('dedicatedServerFtpBackupAccessUpdate', component)
+  .run(/* @ngTranslationsInject:json ./translations */);
+
+export default moduleName;
